backend/requirements.txt
langchain
langchain-community
langchain-huggingface
langchain_groq
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain_huggingface
ipykernel
flask 
flask-cors 
requests